+++ Private/Tui/Read-HydrationTuiEnvironment.ps1
@@ -1,0 +1,12 @@
+function Read-HydrationTuiEnvironment {
+    [CmdletBinding()]
+    [OutputType([string])]
+    param()
+
+    $options = @(Get-HydrationTuiEnvironmentOption)
+    Read-HydrationTuiSingleOption `
+        -Title 'Azure cloud environment' `
+        -Options $options `
+        -Prompt 'Select environment' `
+        -ConvertOption { param($option) $option.Value }
+}